requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
jira>=3.5.2
boto3>=1.28.0
certifi>=2024.2.2
gspread>=5.12.0
google-auth>=2.22.0
openpyxl>=3.1.2
urllib3>=2.0.0
